Perform the shutdown operations in their own thread
The complaint here was that the ShutdownThread inherits the Thread interface and its run method is called directly.
Doing so is bad style and thus this change treats the ShutdownThread class like a thread and executes it in the background waiting for completion - to regain behaviour of the old call semantics.